moved DatePickerView out and times auto populate to next whole hour and hour after that
